Conflicts:
	Administrator/Lib/Administrator.pm

+++ Executor/t/EAddMotherboard.t
@@ -1,9 +1,12 @@
 use Data::Dumper;
-use Log::Log4perl qw(:easy);
+use Log::Log4perl "get_logger";
 use Test::More 'no_plan';
-use Error qw(:try);
 
-use lib qw(../Lib ../../Administrator/Lib);
+use lib qw(../Lib ../../Administrator/Lib ../../Common/Lib);
+
+
+Log::Log4perl->init('../Conf/log.conf');
+my $log = get_logger("executor");
 
 my $admtest = "AdminTest";
 my $exectest = "ExecTest";
@@ -11,26 +14,31 @@ my $exectest = "ExecTest";
 note("Use Tests");
 use_ok(Administrator);
 use_ok(Executor);
+use_ok(McsExceptions);
 
 note("Load Administrator tests");
-Log::Log4perl->easy_init({level=>'DEBUG', file=>'STDOUT', layout=>'%F %L %p %m%n'});
 my @args = ("login",'xebech', "password", 'pass');
 my $adm = new_ok(Administrator => \@args, $admtest);
-
+my $addmotherboard_op;
 note("Operation Addition test");
-my $addmotherboard_op = $adm->newOp(type => "AddMotherboard", priority => '100', params => { mac_address => '00:1c:c0:c0:1c:9a', kernel_id => 2, c_storage_id => 1});
-try {
+
+eval {
+	$addmotherboard_op = $adm->newOp(type => "AddMotherboard", priority => '100', params => { mac_address => '00:1c:c0:c0:1c:9a', kernel_id => 2, c_storage_id => 1});
 	@args = ();
 	my $exec = new_ok("Executor", \@args, $exectest);
 	$exec->execnround(run => 1);
-
+note("Operation Execution is finish");
 	$addmotherboard_op->delete();
-}
-catch Error::Simple with {
-	my $ex = shift;
-	$addmotherboard_op->delete();
-	die "Catch error in Operation instanciation : $ex";
 };
+if ($@){
+	print "Exception catch, its type is : " . ref($@);
+	print Dumper $@;
+	if ($@->isa('Mcs::Exception')) 
+   	{
+		print "Mcs Exception\n";
+   }
+   $addmotherboard_op->delete();
+}
 
 
 #pass($exectest);
